+++ R/zzz.R
@@ -1,7 +1,6 @@
 #' @useDynLib aRtsy
 #' @import Rcpp
 #' @importFrom stats predict
-#' @importFrom dplyr %>%
 NULL
 
 utils::globalVariables(c("x", "y", "z"))