+++ NEWS.md
@@ -7,6 +7,11 @@
 - Added `canvas_cobweb()`.
 - Added `canvas_chladni()`.
 
+**Major changes**
+
+- `canvas_function()` is now able to take a vector of colors for the artwork.
+- Removed the dependency on the `dplyr` package.
+
 # aRtsy 0.1.4
 
 **New features**